# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for the GoChat",
        "title": "GoChat API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/message/channel/{channel_id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Get message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel id",
                        "name": "channel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_message.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Message"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/message/channel/{channel_id}/attachment": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Create attachment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel id",
                        "name": "channel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Attachment data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_message.UploadAttachmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attachment upload data",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.AttachmentUpload"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/message/{message_id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Update message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Message id",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_message.UpdateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Message"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api_endpoints_message.SendMessageRequest": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "cmd_api_endpoints_message.UpdateMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "cmd_api_endpoints_message.UploadAttachmentRequest": {
            "type": "object",
            "properties": {
                "file_size": {
                    "type": "integer"
                },
                "filename": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_FlameInTheDark_gochat_internal_dto.Attachment": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "github_com_FlameInTheDark_gochat_internal_dto.AttachmentUpload": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "type": "integer"
                },
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FlameInTheDark_gochat_internal_dto.Message": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Attachment"
                    }
                },
                "author_id": {
                    "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.User"
                },
                "channel_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "github_com_FlameInTheDark_gochat_internal_dto.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "integer"
                },
                "discriminator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "basic"
        }
    }
}